Refactor margarita-utils
