+++ packages/utils/src/Capitalize/Capitalize.js
