+++ packages/utils/src/CurrencyFormatter/CurrencyFormatter.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { getLocaleDashed } from './GetDeviceLocale';
+import { getLocaleDashed } from '../GetDeviceLocale/GetDeviceLocale';
 
 import 'intl'; // Polyfill because of Android
 import 'intl/locale-data/complete';
