+++ packages/utils/src/Device/Device.js
