+++ packages/utils/src/GetDeviceLocale/GetDeviceLocale.js
