+++ packages/utils/src/GetDeviceLocale/__mocks__/GetDeviceLocale.js
