+++ packages/utils/src/GetDeviceLocale/__mocks__/GetDeviceLocale.native.js
